radio-streaming/requirements.txt
APScheduler==3.6.3
boto3==1.16.4
botocore==1.19.4
click==7.1.2
crontabs==0.2.0
daiquiri==3.0.0
fleming==0.5.0
future==0.18.2
jmespath==0.10.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.3
python-json-logger==2.0.1
pytz==2020.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.20
sqlalchemy-mixins==1.2.1
sqlparse==0.4.1
tabulate==0.8.7
texttable==1.6.3
tzlocal==2.1
urllib3==1.25.11
wcwidth==0.2.5
yoyo-migrations==7.2.0
